+++ RmgLib/src/BaseGrid.cpp
@@ -247,7 +247,7 @@
     }
     size_t BaseGrid::get_GLOBAL_BASIS(int density)
     {
-	return (size_t)(density * density * density) * BaseGrid::NX_GRID * BaseGrid::NY_GRID * BaseGrid::NZ_GRID;
+	return (size_t)(density * density * density) * (size_t)BaseGrid::NX_GRID * (size_t)BaseGrid::NY_GRID * (size_t)BaseGrid::NZ_GRID;
     }
     void BaseGrid::pe2xyz(int pe, int *x, int *y, int *z)
     {
